backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
sqlalchemy>=2.0.25
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
python-dotenv>=1.0.0
aiosmtplib>=3.0.1
email-validator>=2.1.0